+++ src/commands/bot/changelog.js
@@ -4,7 +4,14 @@ module.exports = {
 	description: "Displays the changes made to Phobos.",
 	execute(message) {
 		message.reply({embed: {
-			description: `**May 30**
+			description: `**June 2**
+Added timeToStr utility function
+Added \`rock-paper-scissors\` command - Play rock paper scissors with the bot.
+Enabled partials for CHANNEL, MESSAGE and REACTION
+Changed [\`music.js\`](https://github.com/MarsRon/phobos/blob/master/src/handlers/music.js) to improve readability
+Minor bugfixes
+
+**May 30**
 Added [\`CHANGELOG.md\`](https://github.com/MarsRon/phobos/blob/master/CHANGELOG.md)
 Added a blacklist for [\`wordCatcher.js\`](https://github.com/MarsRon/phobos/blob/master/src/features/wordCatcher.js)
 Removed \`remind\` command because come on who uses it