+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 *This will be very unprofessional please bare with me*
 
+### June 2
+
+ - Added timeToStr utility function
+ - Added `rock-paper-scissors` command - Play rock paper scissors with the bot.
+ - Enabled partials for CHANNEL, MESSAGE and REACTION
+ - Changed [`music.js`](https://github.com/MarsRon/phobos/blob/master/src/handlers/music.js) to improve readability
+ - Minor bugfixes
+
 ### May 30
 
  - Added [`CHANGELOG.md`](https://github.com/MarsRon/phobos/blob/master/CHANGELOG.md)
  - Added a blacklist for [`wordCatcher.js`](https://github.com/MarsRon/phobos/blob/master/src/features/wordCatcher.js)
  - Removed `remind` command because come on who uses it
  - Of course some bug fixes
-
-
 
 ### May 29